Stormifying AverageBolt, fixing bug in WeatherBolt
@ceilingcranes - I commented out the LOGs b/c it all worked. Feel free
to put them back if you want to do any more testing.

+++ src/main/java/fun/connor/storm/WeatherBolt.java
@@ -54,7 +54,7 @@ public class WeatherBolt extends BaseBasicBolt {
                 lon, lat, key);
 
         String regionID = (String) input.getValue(0);
-        Float avgSentiment = (Float) input.getValue(1);
+        Double avgSentiment = (Double) input.getValue(1);
         String tweetID = (String) input.getValue(2);
 
         try {
@@ -78,7 +78,7 @@ public class WeatherBolt extends BaseBasicBolt {
         }
     }
 
-    private String formatOutput(String regionID, Float avgSentiment, String tweetID, JSONObject regionJSON, String weatherJSON) {
+    private String formatOutput(String regionID, Double avgSentiment, String tweetID, JSONObject regionJSON, String weatherJSON) {
         // Parse regionJSON and weatherJSON so they become sub-objects
         Object weatherRaw = null;
         JSONParser parser = new JSONParser();
